change what `get_transforms` returns for VOCDetection
change `transforms.util.get_transforms` so that it returns
the same `target_transform` regardless of the loss function,
when the dataset is `VOCDetection`

it now returns the target transform that converts the class
names in the annotation into a one-hot vector.
This is the target used when training for multi-label.
The `VOCDetection` dataset instantiates other target transforms,
and the `AbstractTrainer` will use the specified `loss_func` to
determine which targets + metrics to use during training.