change loss function used for binary cross entropy
in train, use `BCEWithLogitsLoss` instead of
`BCELoss`. This means the loss function applies
a sigmoid activation itself before computing
the binary cross entropy.

This change is necessary to enable applying other
activations to the output of a single network, for
computing metrics based on different loss functions
while training.
I.e., we want to take the output, compute f1 after
applying sigmoid activation, and then take the same
output and compute accuracy on the largest item
or on a random item.